backend/model/deployement/requirements.txt
model-creator-bird-sing-v2
fastapi
uvicorn[standard]